requirements.txt
pandas
numpy
matplotlib
seaborn
plotly
scikit-learn
imblearn
pymongo
xgboost
catboost
from_root
evidently==0.2.8
dill
PyYAML
neuro_mf
boto3
mypy-boto3-s3
mypy-boto3-secretsmanager
botocore
fastapi
uvicorn
jinja2
python-multipart
-e .